+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-rest/src/test/java/org/activiti/cloud/services/query/rest/ProcessInstanceEntitySearchControllerIT.java
@@ -22,13 +22,9 @@ import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 
 import org.activiti.QueryRestTestApplication;
 import org.activiti.cloud.alfresco.config.AlfrescoWebAutoConfiguration;
-import org.activiti.cloud.services.query.app.repository.ProcessDefinitionRepository;
-import org.activiti.cloud.services.query.app.repository.ProcessInstanceRepository;
-import org.activiti.cloud.services.query.model.ProcessDefinitionEntity;
 import org.activiti.cloud.services.query.model.ProcessInstanceEntity;
 import org.activiti.cloud.services.query.util.ProcessInstanceSearchRequestBuilder;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.http.MediaType;
@@ -51,9 +47,6 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @WithMockUser(username = AbstractProcessInstanceEntitySearchControllerIT.USER, roles = "ACTIVITI_USER")
 class ProcessInstanceEntitySearchControllerIT extends AbstractProcessInstanceEntitySearchControllerIT {
 
-    @Autowired
-    private ProcessDefinitionRepository processDefinitionRepository;
-
     @Container
     @ServiceConnection
     static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:15-alpine");
@@ -61,48 +54,6 @@ class ProcessInstanceEntitySearchControllerIT extends AbstractProcessInstanceEnt
     @Override
     protected String getSearchEndpoint() {
         return "/v1/process-instances/search";
-    }
-
-    @Test
-    public void should_excludeProcessInstances_by_processDefinitionCategoryName() throws Exception {
-        // given
-        ProcessDefinitionEntity processDefinitionToExclude = new ProcessDefinitionEntity();
-        processDefinitionToExclude.setId("proc-def-id-to-exclude");
-        processDefinitionToExclude.setKey("proc-def-key-to-exclude");
-        processDefinitionToExclude.setCategory("CategoryToExclude");
-        processDefinitionRepository.save(processDefinitionToExclude);
-
-        queryTestUtils
-            .buildProcessInstance()
-            .withProcessDefinitionKey(processDefinitionToExclude.getKey())
-            .withInitiator(USER)
-            .buildAndSave();
-
-        ProcessDefinitionEntity processDefinitionToKeep = new ProcessDefinitionEntity();
-        processDefinitionToKeep.setId("proc-def-id-to-keep");
-        processDefinitionToKeep.setKey("proc-def-key-to-keep");
-        processDefinitionToKeep.setCategory("CategoryToKeep");
-        processDefinitionRepository.save(processDefinitionToKeep);
-
-        ProcessInstanceEntity processInstanceToKeep = queryTestUtils
-            .buildProcessInstance()
-            .withProcessDefinitionKey(processDefinitionToKeep.getKey())
-            .withInitiator(USER)
-            .buildAndSave();
-
-        ProcessInstanceSearchRequestBuilder requestBuilder = new ProcessInstanceSearchRequestBuilder()
-            .withExcludeByProcessCategoryName("CategoryToExclude");
-
-        given()
-            .contentType(MediaType.APPLICATION_JSON)
-            .body(requestBuilder.buildJson())
-            .when()
-            .post(getSearchEndpoint())
-            // then
-            .then()
-            .statusCode(200)
-            .body("_embedded.processInstances", hasSize(1))
-            .body("_embedded.processInstances[0].id", equalTo(processInstanceToKeep.getId()));
     }
 
     @Test
